requirements.txt
Django==5.2.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
pandas==2.2.2
openpyxl==3.1.5
django-cors-headers
openpyxl